==> 2_Analysis_bx_mafq.Rmd <==
---
title: "2_Analysis_bx_mafq"
author: "Jackie Kirshenbaum"
date: "`r format(Sys.Date(), '%B %e %Y')`"
output:
  html_document:
    df_print: paged
  toc: yes
  toc_float: yes
editor_options: 
  chunk_output_type: console
---

# Libraries 

```{r, libraries, message=FALSE}
library(tidyverse)
library(foreign)
library(readxl)
library(lubridate)
library(lmerTest)
library(psych)
library(viridis)
library(hrbrthemes)
library(forcats)
library(DescTools)
library(rstatix)
library(car)
library(ggpubr)
library(robustHD)
library(qwraps2)
library(knitr)
library(lme4)
```


# Read in Data

```{r, filepaths, message=FALSE}
# pad data
pad_data_fp <- "~/Jackie/pad_data_2021_05_11.Rda"
pad_data <- load(pad_data_fp)

# roi data from 1_Analysis_brain_and_sample_characteristics.Rmd
rois_pad_fp <- "~/Jackie/rois_and_pad_usable.csv"
rois_pad <- 
  read_csv(rois_pad_fp)

```

## Merging brain and long pad data

```{r}
pad_data2 <-
  pad_data_long %>%
  dplyr::mutate(subID = factor(subID)) %>%
  drop_na(mafq_c)

braindata <-
  rois_pad %>%
  dplyr::select(subID, brain_data) %>%
  distinct() %>%
  dplyr::mutate(subID = factor(subID))

rois2 <-
  rois_pad %>%
  dplyr::select(subID, condition, starts_with("AntIns"), starts_with("dACC"), starts_with("NAcc"), starts_with("caudate"), starts_with("putamen")) %>%
  dplyr::mutate(subID = factor(subID))

pad_data_hasbrain <-
  left_join(
    pad_data2,
    braindata,
    by = "subID"
  ) %>%
  drop_na(brain_data) %>%
  dplyr::select(
    subID, visit, visitN, N_Long, Group, Age, TannerMean, Sex, Hispanic, Race, MaritalStatus, ParentEducation, OtherParentEducation, FamilyIncome, ParentEmploymentIncome, ChildSupport, FederalAid, Disability, White, ParentsMarried, FinancialAssistance, aleq_c:aleqromantic_c, starts_with("mafq"), masc_c, starts_with("masq"), shaps_c, shapscount_c 
  ) %>%
  dplyr::mutate(subID = factor(subID))

```


# Distribution of mafq across time

_For those with brain data_

```{r}
pad_data_hasbrain %>%
  ggplot(
    aes(x = visit, y = mafq_c)
  ) +
  geom_line(aes(color = subID)) +
  geom_smooth(method = "loess", se = FALSE, color = "black") +
  theme_classic() +
  theme(legend.position = "none") +
  labs(
    x = "Visit",
    y = "Depression Severity",
    title = "Depression Severity over 2 Years"
  ) # on average, no change in depressive sx severity over time (which is expected), but there is a lot of variability within and between persons
```

> looks like 2 outlying depression scores. Checking skew and kurtosis for each visit


```{r}
library(ltm)
library(moments)
# distribution of mafq at baseline
pad_and_rois_v0 <-
  pad_data_hasbrain %>%
  filter(visit == 0)
pad_and_rois_v0 %>%
  ggplot(
    aes(x = mafq_c)
  ) +
  geom_histogram() +
  theme_classic()
skewness(pad_and_rois_v0$mafq_c)
kurtosis(pad_and_rois_v0$mafq_c)

# distribution of mafq at fu 1
pad_and_rois_v1 <-
  pad_data_hasbrain %>%
  filter(visit == 1)
pad_and_rois_v1 %>%
  ggplot(
    aes(x = mafq_c)
  ) +
  geom_histogram() +
  theme_classic()
skewness(pad_and_rois_v1$mafq_c, na.rm = TRUE) 
kurtosis(pad_and_rois_v1$mafq_c, na.rm = TRUE) 

# distribution of mafq at fu 3
pad_and_rois_v3 <-
  pad_data_hasbrain %>%
  filter(visit == 3)
pad_and_rois_v3 %>%
  ggplot(
    aes(x = mafq_c)
  ) +
  geom_histogram() +
  theme_classic()
skewness(pad_and_rois_v3$mafq_c, na.rm = TRUE) 
kurtosis(pad_and_rois_v3$mafq_c, na.rm = TRUE) 

# distribution of mafq at fu 6
pad_and_rois_v6 <-
  pad_data_hasbrain %>%
  filter(visit == 6)
pad_and_rois_v6 %>%
  ggplot(
    aes(x = mafq_c)
  ) +
  geom_histogram() +
  theme_classic()
skewness(pad_and_rois_v6$mafq_c, na.rm = TRUE) 
kurtosis(pad_and_rois_v6$mafq_c, na.rm = TRUE) 

# distribution of mafq at fu 9
pad_and_rois_v9 <-
  pad_data_hasbrain %>%
  filter(visit == 9)
pad_and_rois_v9 %>%
  ggplot(
    aes(x = mafq_c)
  ) +
  geom_histogram() +
  theme_classic()
skewness(pad_and_rois_v9$mafq_c, na.rm = TRUE) 
kurtosis(pad_and_rois_v9$mafq_c, na.rm = TRUE) 

# distribution of mafq at fu 12
pad_and_rois_v12 <-
  pad_data_hasbrain %>%
  filter(visit == 12)
pad_and_rois_v12 %>%
  ggplot(
    aes(x = mafq_c)
  ) +
  geom_histogram() +
  theme_classic()
skewness(pad_and_rois_v12$mafq_c, na.rm = TRUE) 
kurtosis(pad_and_rois_v12$mafq_c, na.rm = TRUE) 

# distribution of mafq at fu 15
pad_and_rois_v15 <-
  pad_data_hasbrain %>%
  filter(visit == 15)
pad_and_rois_v15 %>%
  ggplot(
    aes(x = mafq_c)
  ) +
  geom_histogram() +
  theme_classic()
skewness(pad_and_rois_v15$mafq_c, na.rm = TRUE)
kurtosis(pad_and_rois_v15$mafq_c, na.rm = TRUE)  

# distribution of mafq at fu 18
pad_and_rois_v18 <-
  pad_data_hasbrain %>%
  filter(visit == 18)
pad_and_rois_v18 %>%
  ggplot(
    aes(x = mafq_c)
  ) +
  geom_histogram() +
  theme_classic()
skewness(pad_and_rois_v18$mafq_c, na.rm = TRUE) 
kurtosis(pad_and_rois_v18$mafq_c, na.rm = TRUE) 

# distribution of mafq at fu 21
pad_and_rois_v21 <-
  pad_data_hasbrain %>%
  filter(visit == 21)
pad_and_rois_v21 %>%
  ggplot(
    aes(x = mafq_c)
  ) +
  geom_histogram() +
  theme_classic()
skewness(pad_and_rois_v21$mafq_c, na.rm = TRUE) 
kurtosis(pad_and_rois_v21$mafq_c, na.rm = TRUE) 

# distribution of mafq at fu 24
pad_and_rois_v24 <-
  pad_data_hasbrain %>%
  filter(visit == 24)
pad_and_rois_v24 %>%
  ggplot(
    aes(x = mafq_c)
  ) +
  geom_histogram() +
  theme_classic()
skewness(pad_and_rois_v24$mafq_c, na.rm = TRUE) 
kurtosis(pad_and_rois_v24$mafq_c, na.rm = TRUE) 
```

> based on examination of each visit, kurtosis is pretty bad. Going to winsorize

## Winsorizing

```{r}
library(DescTools)
# identifying outliers
pad_and_rois_outliers <-
  pad_data_hasbrain %>%
  ungroup() %>%
  identify_outliers(mafq_c) %>%
  filter(is.extreme == TRUE) # 12

sum_mafq <-
  pad_data_hasbrain %>%
  ungroup() %>%
  dplyr::summarise(
    sdmafq_c = sd(mafq_c, na.rm = TRUE),
    meanmafq_c = mean(mafq_c, na.rm = TRUE)
  )

Mplus3SD <- sum_mafq$meanmafq_c[1] + (3 * sum_mafq$sdmafq_c[1])
print(Mplus3SD)
Mminus3SD <- sum_mafq$meanmafq_c[1] - (3 * sum_mafq$sdmafq_c[1])
print(Mplus3SD)

pad_and_rois_w <-
  pad_data_hasbrain %>%
  dplyr::ungroup() %>%
  dplyr::mutate(
    mafq_c_w = Winsorize(mafq_c, minval = Mminus3SD, maxval = Mplus3SD, probs = c(0.05, 0.95)),
    mafq_c_w = round(mafq_c_w, 1),
    match_mafq_c_w = mafq_c == mafq_c_w
  )

### how much of the data were winsorized
perc_win <-
  pad_and_rois_w %>%
  group_by(match_mafq_c_w) %>%
  dplyr::summarise(
    n = n(),
    perc = n/656
  ) # ~ 2% of data winsorized with 3sd


# REidentifying outliers
roi_long_outliers <-
  pad_and_rois_w %>%
  identify_outliers(mafq_c_w) %>%
  filter(is.extreme == TRUE) # 0  outliers

skewness(pad_and_rois_w$mafq_c_w, na.rm = TRUE) # still bad
pad_and_rois_w <-
  pad_and_rois_w %>%
  dplyr::mutate(
    mafq_c_w_log = log10(mafq_c_w),
    mafq_c_w_sqrt = sqrt(mafq_c_w)
  )
skewness(pad_and_rois_w$mafq_c_w_log) # 0.01581427 better
skewness(pad_and_rois_w$mafq_c_w_sqrt) # 0.7849932 not as good

pad_and_rois_w %>%
  ggplot(
    aes(x = visit, y = mafq_c_w_log)
  ) +
  geom_line(aes(color = subID)) +
  geom_smooth(method = "loess", se = FALSE, color = "black") +
  scale_y_continuous(limits = c(0, 2)) +
  theme_classic() +
  theme(legend.position = "none") +
  labs(
    x = "Visit",
    y = "Depression Severity",
    title = "Depression Severity over 2 Years"
  )  # slightly better
```

```{r}
pad_and_rois_w <-
  pad_and_rois_w %>%
  mutate(subID = factor(subID))

```


# Plotting overall distributions of depression and stress

```{r}
pad_and_rois_w %>%
  ggplot(aes(x=mafq_c_w)) +
  geom_histogram(fill = "white", color = "black") +
  labs(x = "Depression Severity")

pad_and_rois_w %>%
  ggplot(aes(x=aleq_c)) +
  geom_histogram(fill = "white", color = "black") +
  labs(x = "Stress Severity")
```

# Means and SDs for MAFQ

```{r}
means_by_visit <-
  pad_and_rois_w %>%
  dplyr::group_by(visit) %>%
  dplyr::summarise(
    mean_mafq = round(mean(mafq_c_w), 2),
    sd_mafq = round(sd(mafq_c_w), 2)
  )
write_csv(means_by_visit, "mafq_means_by_visit.csv")
```

# How do depression and stress go together? 

## Intraindividual plots (concurrent)
```{r}
i <- c("sub-HP603", "sub-HP800", "sub-HP823", "sub-LP205", "sub-LP208", "sub-LP426")
pad_and_rois_w %>%
  filter(subID %in% i) %>%
  ggplot(
    aes(x=visit, group=subID), legend=FALSE) +
  #geom_rect(mapping=aes(xmin=day-.5, xmax=day+.5, ymin=0, ymax=10, fill=wrkstrscw), alpha=0.6) +
  geom_point(aes(x=visitN,y = scale(mafq_c_w_log)), color="blue", shape=17, size=2) +
  geom_line(aes(x=visitN,y = scale(mafq_c_w_log)), color="blue", lty=1, size=1) +
  geom_point(aes(x=visitN,y = scale(aleq_c)), color="red", shape=17, size=2) +
  geom_line(aes(x=visitN,y = scale(aleq_c)), color="red", lty=1, size=1) +
  xlab("Visit") + 
  ylab("Depression = Blue, Stress = Red") +
  scale_x_continuous(breaks=seq(0,9,by=1)) +
  facet_wrap( ~ subID) +
  theme_classic()
```
## But, the variance can comes from between and within-persons

Level 1 variable centering
- computing deviations from the person's mean of that variable (x - mean(x)) for that person
Level 2 variable centering
- the person's mean (mean(x))

### First, we split our predictor variables into time-varying and time-invariant components
```{r}
# ordering by subject and by visit
pad_and_rois_w_order <- pad_and_rois_w[order(pad_and_rois_w$subID, pad_and_rois_w$visitN),]

# combining peer and family stress to create "interpersonal stress"
interpersonalvars <- c("aleqpeer_c", "aleqfamily_c")

pad_and_rois_w_order$aleqinterpersonal_c = rowSums(pad_and_rois_w_order[,(interpersonalvars)])
```

```{r}
library(plyr)
library(dplyr)
library(reshape2)
visit.means_aleq <- ddply(pad_and_rois_w_order, "subID", summarize, aleq_c.L2 =mean(aleq_c, na.rm=TRUE))
visit.means_aleqNonInt <- ddply(pad_and_rois_w_order, "subID", summarize, aleqNonInt_c.L2 =mean(aleqnoninterpersonal_c, na.rm=TRUE))
visit.means_aleqInt <- ddply(pad_and_rois_w_order, "subID", summarize, aleqInt_c.L2 =mean(aleqinterpersonal_c, na.rm=TRUE))

#Calculating sample-centered versions *Note that this is done in a person-level data file.
visit.means_aleq$aleq_c.L2.c <- scale(visit.means_aleq$aleq_c.L2,center=TRUE,scale=FALSE)
visit.means_aleqNonInt$aleqNonInt_c.L2.c <- scale(visit.means_aleqNonInt$aleqNonInt_c.L2, center = TRUE, scale = FALSE)
visit.means_aleqInt$aleqInt_c.L2.c <- scale(visit.means_aleqInt$aleqInt_c.L2, center = TRUE, scale = FALSE)



#merging "trait" scores back into the *long* data file and calculate "state" scores.
visit.means2 <- merge(
  pad_and_rois_w_order,
  visit.means_aleq,
  by="subID"
  )
visit.means3 <- merge(
  visit.means2,
  visit.means_aleqNonInt,
  by="subID"
  )
visit.means3 <- merge(
  visit.means3,
  visit.means_aleqInt,
  by="subID"
  )

# creating level 1 variables
visit.means3$aleq_c.L1 <- visit.means3$aleq_c - visit.means3$aleq_c.L2
visit.means3$aleqNonInt_c.L1 <- visit.means3$aleqnoninterpersonal_c - visit.means3$aleqNonInt_c.L2
visit.means3$aleqInt_c.L1 <- visit.means3$aleqinterpersonal_c - visit.means3$aleqInt_c.L2

```

> L2 = level 2 (trait/btwn-person) variance
> L1 = level 1 (state/within-person) variance

### saving L1 and L2 variable dataframe 
```{r}
visit.means3_L1andL2vars <-
  visit.means3 %>%
  dplyr::select(-ends_with("L2.c"))
```

# ACF
```{r}
library(forecast)
Acf(visit.means3_L1andL2vars$mafq_c_w, type = "correlation", plot = TRUE, na.action = na.contiguous)
# mostly surrounding zero
```

## Regular MLMs using within and between person variance of stress to predict trajectories of depression
```{r,message=FALSE}
library(sjPlot)
library(lmerTest)
library(performance)
library(reghelper)
library(parameters)
library(MuMIn)
library(interactions)
```

# Bx Model 
> To determine covariates

```{r}
contrasts(visit.means3_L1andL2vars$Sex)
contrasts(visit.means3_L1andL2vars$Group)

mlm_ri <- 
  lmer(
    scale(mafq_c_w)  ~  
      scale(aleq_c.L1) +  scale(aleq_c.L2)  + scale(visitN) + scale(Age) + Sex + scale(TannerMean) + Group +
      (1 | subID), REML = TRUE,
    data = visit.means3_L1andL2vars
    )
summary(mlm_ri) # Tanner, Age, between and within-person stress


mlm_rs <- 
  lmer(
    scale(mafq_c_w)  ~  
      scale(aleq_c.L2)  + scale(aleq_c.L1)  + scale(visitN) + scale(Age) + Sex + scale(TannerMean) + Group +
      (1 + scale(visitN)| subID), REML = TRUE,
    data = visit.means3_L1andL2vars
    )
summary(mlm_rs) # Tanner, Age, between and within-person stress; # No interaction btwn aleq_c.L1 * visitN (so, stress does not vary linearly with time)

anova(mlm_ri, mlm_rs) # random slope is better fit 

check_model(mlm_rs, type = "diag")
```
> going with random slope model 

```{r}
sjPlot::tab_model(mlm_rs, bootstrap = TRUE, iterations = 1000, dv.labels = c("Depression Severity"), seed = 1234, pred.labels = c("Intercept", "Between-person Stress", "Within-person Stress", "Visit", "Age", "Sex [F]", "Tanner", "Risk Group [High]"), file = "Depression_Covariates.doc")
```

# Writing csv for next analysis
```{r}

pad_and_rois_usable <- 
  left_join(
    visit.means3_L1andL2vars,
    rois2,
    by = "subID"
  )

write_csv(pad_and_rois_usable, "pad_and_rois_usable_mafq.csv")
```
