==> 2_Correlations_Bx.Rmd <==
---
title: "2_Corrleations_Bx"
author: "Jackie Kirshenbaum"
date: "`r format(Sys.Date(), '%B %e %Y')`"
output:
  html_document:
    df_print: paged
  toc: yes
  toc_float: yes
---

Correlations among depression, anhedonia, and anxiety

# Libraries 

```{r, libraries, message=FALSE}
library(tidyverse)
library(foreign)
library(readxl)
library(psych)
library(viridis)
library(hrbrthemes)
library(ggpubr)
library(rstatix)
```


# Read in Data

```{r, filepaths, message=FALSE}
# pad data
pad_data_fp <- "~/Jackie/pad_data_2021_05_11.Rda"
pad_data <- load(pad_data_fp)

# roi data from 1_Analysis_brain_and_sample_characteristics.Rmd
rois_pad_fp <- "~/Jackie/rois_and_pad_usable.csv"
rois_pad <- 
  read_csv(rois_pad_fp)

```

## Merging brain and long pad data

```{r}
pad_data2 <-
  pad_data_long %>%
  dplyr::mutate(subID = factor(subID)) %>%
  drop_na(masc_c)

braindata <-
  rois_pad %>%
  dplyr::select(subID, brain_data) %>%
  distinct() %>%
  dplyr::mutate(subID = factor(subID))

rois2 <-
  rois_pad %>%
  dplyr::select(subID, condition, starts_with("AntIns"), starts_with("dACC"), starts_with("NAcc"), starts_with("caudate"), starts_with("putamen")) %>%
  dplyr::mutate(subID = factor(subID))

pad_data_hasbrain <-
  left_join(
    pad_data2,
    braindata,
    by = "subID"
  ) %>%
  drop_na(brain_data) %>%
  dplyr::select(
    subID, visit, visitN, N_Long, Group, Age, TannerMean, Sex, Hispanic, Race, MaritalStatus, ParentEducation, OtherParentEducation, FamilyIncome, ParentEmploymentIncome, ChildSupport, FederalAid, Disability, White, ParentsMarried, FinancialAssistance, aleq_c:aleqromantic_c, mafq_c, masc_c, masqharmavoid_c, masqphysical_c, masqsepanx_c, masqsocial_c, shaps_c, shapscount_c 
  ) %>%
  dplyr::mutate(
    masc_c_total = masqharmavoid_c + masqphysical_c + masqsepanx_c + masqsocial_c,
    aleqinterpersonal_c = aleqfamily_c + aleqpeer_c + aleqromantic_c,
    subID = factor(subID)
    )

```

## How many follow-ups were completed?

```{r}
getmode <- function(v) {
   uniqv <- unique(v)
   uniqv[which.max(tabulate(match(v, uniqv)))]
}

visit_desc <-
  pad_data_hasbrain %>%
  filter(visit == "0") %>%
  dplyr::select(subID, N_Long) # participants completed an average of 8 follow ups

visit_desc %>%
  dplyr::summarise(
    visit_num_avg = mean(N_Long),
    visit_num_mode = getmode(N_Long)
  ) # most participants completed 10 follow-ups

visit_desc %>%
  dplyr::mutate(N_Long = factor(N_Long)) %>%
  ggplot(
    aes(x = N_Long, fill = N_Long)
  ) +
  geom_histogram(stat = "count")


```

### Reading in winsorized mafq socres
_This was the only clinical measure that needed to be winsorized_
```{r}
mafq_w_df <- 
  read_csv("pad_and_rois_usable_mafq.csv") %>%
  dplyr::select(subID, mafq_c_w, visit, visitN) %>%
  mutate(
    subID = factor(subID)
  )

### merging with rest of df
pad_data_hasbrain2 <-
  left_join(
    pad_data_hasbrain,
    mafq_w_df,
    by = c("subID", "visit", "visitN")
  )

```

# Distribution of mafq across time

_For those with brain data_

```{r}
pad_data_hasbrain2 %>%
  ggplot(
    aes(x = visit, y = mafq_c_w)
  ) +
  geom_line(aes(color = subID)) +
  geom_smooth(method = "loess", se = FALSE, color = "black") +
  theme_classic() +
  theme(legend.position = "none") +
  labs(
    x = "Visit",
    y = "Depression Severity",
    title = "Depression Severity over 2 Years"
  ) # on average, no change in depressive sx severity over time (which is expected), but there is a lot of variability within and between persons


pad_data_hasbrain2 %>%
  ggplot(aes(x=mafq_c_w)) +
  geom_histogram(color = "black") +
  labs(x = "Depression Severity")

mean(pad_data_hasbrain2$mafq_c_w)
sd(pad_data_hasbrain2$mafq_c_w)
summary(pad_data_hasbrain2$mafq_c_w)
```

# Distribution of shaps across time

_For those with brain data_

```{r}
pad_data_hasbrain2 %>%
  ggplot(
    aes(x = visit, y = shaps_c)
  ) +
  geom_line(aes(color = subID)) +
  geom_smooth(method = "loess", se = FALSE, color = "black") +
  theme_classic() +
  theme(legend.position = "none") +
  labs(
    x = "Visit",
    y = "Anhedonia Severity",
    title = "Anhedonia Severity over 2 Years"
  ) # on average, no change in anhedonia sx severity over time (which is expected), but there is a lot of variability within and between persons

pad_data_hasbrain2 %>%
  ggplot(aes(x=shaps_c)) +
  geom_histogram(color = "black") +
  labs(x = "Anhedonia Severity")

mean(pad_data_hasbrain2$shaps_c)
sd(pad_data_hasbrain2$shaps_c)
summary(pad_data_hasbrain2$shaps_c)
```


# Distribution of masc across time

_For those with brain data_

```{r}
pad_data_hasbrain2 %>%
  ggplot(
    aes(x = visit, y = masc_c_total)
  ) +
  geom_line(aes(color = subID)) +
  geom_smooth(method = "loess", se = FALSE, color = "black") +
  theme_classic() +
  theme(legend.position = "none") +
  labs(
    x = "Visit",
    y = "Anxiety Severity",
    title = "Anxiety Severity over 2 Years"
  ) # on average, no change in anxiety sx severity over time (which is expected), but there is a lot of variability within and between persons

pad_data_hasbrain2 %>%
  ggplot(aes(x=masc_c)) +
  geom_histogram(color = "black") +
  labs(x = "Anxiety Severity")

mean(pad_data_hasbrain$masc_c_total)
sd(pad_data_hasbrain$masc_c_total)
summary(pad_data_hasbrain$masc_c_total)
```

# How do depression and stress go together? 

## Intraindividual plots (concurrent)
```{r}
i <- c("sub-HP603", "sub-HP800", "sub-HP823", "sub-LP205", "sub-LP208", "sub-LP426")
pad_data_hasbrain2 %>%
  filter(subID %in% i) %>%
  ggplot(
    aes(x=visit, group=subID), legend=FALSE) +
  #geom_rect(mapping=aes(xmin=day-.5, xmax=day+.5, ymin=0, ymax=10, fill=wrkstrscw), alpha=0.6) +
  geom_point(aes(x=visitN,y = scale(mafq_c_w)), color="blue", shape=17, size=2) +
  geom_line(aes(x=visitN,y = scale(mafq_c_w)), color="blue", lty=1, size=1) +
  geom_point(aes(x=visitN,y = scale(aleq_c)), color="red", shape=17, size=2) +
  geom_line(aes(x=visitN,y = scale(aleq_c)), color="red", lty=1, size=1) +
  xlab("Visit") + 
  ylab("Depression = Blue, Stress = Red") +
  scale_x_continuous(breaks=seq(0,9,by=1)) +
  facet_wrap( ~ subID) +
  theme_classic()
```

# How do anhedonia and stress go together? 

## Intraindividual plots (concurrent)
```{r}
i <- c("sub-HP603", "sub-HP800", "sub-HP823", "sub-LP205", "sub-LP208", "sub-LP426")
pad_data_hasbrain2 %>%
  filter(subID %in% i) %>%
  ggplot(
    aes(x=visit, group=subID), legend=FALSE) +
  #geom_rect(mapping=aes(xmin=day-.5, xmax=day+.5, ymin=0, ymax=10, fill=wrkstrscw), alpha=0.6) +
  geom_point(aes(x=visitN,y = scale(shaps_c)), color="chartreuse4", shape=17, size=2) +
  geom_line(aes(x=visitN,y = scale(shaps_c)), color="chartreuse4", lty=1, size=1) +
  geom_point(aes(x=visitN,y = scale(aleq_c)), color="red", shape=17, size=2) +
  geom_line(aes(x=visitN,y = scale(aleq_c)), color="red", lty=1, size=1) +
  xlab("Visit") + 
  ylab("Anhedonia = Green, Stress = Red") +
  scale_x_continuous(breaks=seq(0,9,by=1)) +
  facet_wrap( ~ subID) +
  theme_classic()
```

# How do anxiety and stress go together? 

## Intraindividual plots (concurrent)
```{r}
i <- c("sub-HP603", "sub-HP800", "sub-HP823", "sub-LP205", "sub-LP208", "sub-LP426")
pad_data_hasbrain2 %>%
  filter(subID %in% i) %>%
  ggplot(
    aes(x=visit, group=subID), legend=FALSE) +
  #geom_rect(mapping=aes(xmin=day-.5, xmax=day+.5, ymin=0, ymax=10, fill=wrkstrscw), alpha=0.6) +
  geom_point(aes(x=visitN,y = scale(masc_c_total)), color="darkorange", shape=17, size=2) +
  geom_line(aes(x=visitN,y = scale(masc_c_total)), color="darkorange", lty=1, size=1) +
  geom_point(aes(x=visitN,y = scale(aleq_c)), color="red", shape=17, size=2) +
  geom_line(aes(x=visitN,y = scale(aleq_c)), color="red", lty=1, size=1) +
  xlab("Visit") + 
  ylab("Anxiety = Orange, Stress = Red") +
  scale_x_continuous(breaks=seq(0,9,by=1)) +
  facet_wrap( ~ subID) +
  theme_classic()
```



```{r}
library(scipub)
library(corrr)

bx_corrtable <-
  correltable(
    pad_data_hasbrain2,
    vars = c("mafq_c_w", "shaps_c", "masc_c_total", "Age", "TannerMean"),
    var_names = c("Depression", "Anhedonia", "Anxiety", "Age (years)", "Tanner Average"),
    tri = "lower"
  )
bx_corrtable

bx_corrtableSex <-
  correltable(
    pad_data_hasbrain2,
    vars = c("mafq_c_w", "shaps_c", "masc_c_total", "Age", "TannerMean"),
    var_names = c("Depression", "Anhedonia", "Anxiety", "Age (years)", "Tanner Average"),
    strata = "Sex"
  )
bx_corrtableSex

bx_corrtableRisk <-
  correltable(
    pad_data_hasbrain2,
    vars = c("mafq_c_w", "shaps_c", "masc_c_total", "Age", "TannerMean"),
    var_names = c("Depression", "Anhedonia", "Anxiety", "Age (years)", "Tanner Average"),
    strata = "Group"
  )


## RM Corr
library(rmcorr)
mfq_shaps <-
  rmcorr::rmcorr(
    subID,
    mafq_c_w,
    shaps_c,
    dataset = pad_data_hasbrain2
  )
mfq_shaps
plot(mfq_shaps)

mfq_masc <-
  rmcorr::rmcorr(
    subID,
    mafq_c_w,
    masc_c_total,
    dataset = pad_data_hasbrain2
  )
mfq_masc
plot(mfq_masc)

mfq_aleq <-
  rmcorr::rmcorr(
    subID,
    mafq_c_w,
    aleq_c,
    dataset = pad_data_hasbrain2
  )
mfq_aleq
plot(mfq_aleq)

mfq_aleqInt <-
  rmcorr::rmcorr(
    subID,
    mafq_c_w,
    aleqinterpersonal_c,
    dataset = pad_data_hasbrain2
  )
mfq_aleqInt
plot(mfq_aleqInt)

mfq_aleqNonInt <-
  rmcorr::rmcorr(
    subID,
    mafq_c_w,
    aleqnoninterpersonal_c,
    dataset = pad_data_hasbrain2
  )
mfq_aleqNonInt
plot(mfq_aleqNonInt)

shaps_masc <-
  rmcorr::rmcorr(
    subID,
    shaps_c,
    masc_c_total,
    dataset = pad_data_hasbrain2
  )
shaps_masc
plot(shaps_masc)

shaps_aleq <-
  rmcorr::rmcorr(
    subID,
    shaps_c,
    aleq_c,
    dataset = pad_data_hasbrain2
  )
shaps_aleq
plot(shaps_aleq)

shaps_aleqInt <-
  rmcorr::rmcorr(
    subID,
    shaps_c,
    aleqinterpersonal_c,
    dataset = pad_data_hasbrain2
  )
shaps_aleqInt
plot(shaps_aleqInt)

shaps_aleqNonInt <-
  rmcorr::rmcorr(
    subID,
    shaps_c,
    aleqnoninterpersonal_c,
    dataset = pad_data_hasbrain2
  )
shaps_aleqNonInt
plot(shaps_aleqInt)

masc_aleq <-
  rmcorr::rmcorr(
    subID,
    masc_c_total,
    aleq_c,
    dataset = pad_data_hasbrain2
  )
masc_aleq
plot(masc_aleq)

masc_aleqInt <-
  rmcorr::rmcorr(
    subID,
    masc_c_total,
    aleqinterpersonal_c,
    dataset = pad_data_hasbrain2
  )
masc_aleqInt
plot(masc_aleqInt)

masc_aleqNonInt <-
  rmcorr::rmcorr(
    subID,
    masc_c_total,
    aleqnoninterpersonal_c,
    dataset = pad_data_hasbrain2
  )
masc_aleqNonInt
plot(masc_aleqNonInt)

aleq_aleqInt <-
  rmcorr::rmcorr(
    subID,
    aleq_c,
    aleqinterpersonal_c,
    dataset = pad_data_hasbrain2
  )
aleq_aleqInt
plot(aleq_aleqInt)

aleq_aleqNonInt <-
  rmcorr::rmcorr(
    subID,
    aleq_c,
    aleqnoninterpersonal_c,
    dataset = pad_data_hasbrain2
  )
aleq_aleqNonInt
plot(aleq_aleqNonInt)

aleqInt_NonInt <-
  rmcorr::rmcorr(
    subID,
    aleqinterpersonal_c,
    aleqnoninterpersonal_c,
    dataset = pad_data_hasbrain2
  )
aleqInt_NonInt
plot(aleqInt_NonInt)
```

## VAS RATINGS PRE ##
```{r}
library(readxl)
prefp <- "~/Behavioral Tasks/Adolescent/Day 2/Affect_PRE/PRE_MERGE.xlsx"
# Read in an Eprime text file
pre_vas <- read_excel(prefp, sheet = "VAS_PRE_RATINGS")

pre_vas2 <-
  pre_vas %>%
  dplyr::rename("subID" = Subject) %>%
  dplyr::rename("PreRating" = XRatingFinal) %>%
  dplyr::mutate(subID = as.factor(subID)) %>%
  dplyr::mutate(word = as.factor(word)) %>%
  pivot_wider(names_from = word, values_from = PreRating) %>%
  rowwise() %>%
  dplyr::mutate(
    Pre_PA = mean(c(HAPPY, JOYFUL)),
    Pre_NA = mean(c(UPSET, SAD, DISCOURAGED))
    ) %>%
  dplyr::select(subID, Pre_PA, Pre_NA)
```

## VAS RATINGS POST ##
```{r}
library(readxl)
postfp <- "~/Behavioral Tasks/Adolescent/Day 2/Affect_POST/POST_MERGE.xlsx"
# Read in an Eprime text file
post_vas <- read_excel(postfp, sheet = "VAS_POST_RATINGS")

post_vas2 <-
  post_vas %>%
  dplyr::rename("subID" = Subject) %>%
  dplyr::rename("PostRating" = XRatingFinal) %>%
  dplyr::mutate(subID = as.factor(subID)) %>%
  dplyr::mutate(word = as.factor(word)) %>%
  pivot_wider(names_from = word, values_from = PostRating) %>%
  rowwise() %>%
  dplyr::mutate(
    Post_PA = mean(c(HAPPY, JOYFUL)),
    Post_NA = mean(c(UPSET, SAD, DISCOURAGED))
    )  %>%
  dplyr::select(subID, Post_PA, Post_NA)
```

### MERGE PRE AND POST VAS


```{r}
vas <-
  left_join(pre_vas2, post_vas2, by = "subID")

braindata_subj_variable_remove_label <-
  braindata %>%
  dplyr::mutate(subID = as.character(subID)) %>%
  dplyr::mutate(
    subID = str_remove(subID, "sub-HP"),
    subID = str_remove(subID, "sub-LP")
  ) %>%
  dplyr::mutate(subID = as.factor(subID))

vas_usable <- 
  left_join(braindata_subj_variable_remove_label, vas, by = "subID")

vas_long <-
  vas_usable %>%
  pivot_longer(
    cols = c(starts_with("Pre"), starts_with("Post")),
    names_to = "Run"
  ) %>%
  # mutate(
  #   across(subID:rois, factor),
  #   value = as.numeric(value)
  #   )%>%
  separate(., col = "Run", into = c("run", "affect"), sep = "([_])") %>%
  dplyr::mutate(run = factor(run))

# PA
vas_long_PA <-
  vas_long %>%
  filter(affect == "PA")

t.test(value ~ run, data = vas_long_PA)

# effect size PA change
vas_wide_PA <- vas_long_PA %>% pivot_wider(names_from = run, values_from = value)
pre_pa_mean <- mean(vas_wide_PA$Pre, na.rm =TRUE)
post_pa_mean <- mean(vas_wide_PA$Post, na.rm =TRUE)
vas_pa_cohensd <- 
  (pre_pa_mean - post_pa_mean)/sqrt(((var(vas_wide_PA$Pre, na.rm = TRUE)) + (var(vas_wide_PA$Post, na.rm = TRUE)))/2) # 1.424192

# NA
vas_long_NA <-
  vas_long %>%
  filter(affect == "NA")

t.test(value ~ run, data = vas_long_NA)

# effect size NA change
vas_wide_NA <- vas_long_NA %>% pivot_wider(names_from = run, values_from = value)
pre_na_mean <- mean(vas_wide_NA$Pre, na.rm =TRUE)
post_na_mean <- mean(vas_wide_NA$Post, na.rm =TRUE)
vas_na_cohensd <- 
  (post_na_mean - pre_na_mean)/sqrt(((var(vas_wide_NA$Post, na.rm = TRUE)) + (var(vas_wide_NA$Pre, na.rm = TRUE)))/2) # 1.698194

vas_long$run = fct_relevel(vas_long$run, "Pre", "Post")

# computing 95% CI
affectsum <-
  vas_long %>%
  dplyr::group_by(run, affect) %>%
  dplyr::summarise(
    n = n(),
    mean = mean(value, na.rm = TRUE),
    sd = sd(value, na.rm = TRUE)
  ) %>%
  mutate(se=sd/sqrt(n))  %>%
  mutate(ic=se * qt((1-0.05)/2 + .5, n-1))
limits <- aes(ymin=affectsum$mean-affectsum$ic, ymax=affectsum$mean+affectsum$ic)

stat.test <- compare_means(
  value ~ run, data = vas_long, group.by = "affect",
  method = "t.test"
)

vas_long %>%
  dplyr::mutate(
    run_type = recode_factor(run, Pre = "Pre-Stress", Post = "Post-Stress")
    ) %>%
  ggbarplot(
    x = "run_type", y = "value", fill = "run_type",
    add = c("mean_ci"),
    add.params = list(color = "black", size = 1),
    position = position_dodge(),
    width = 1,
    palette = c("darkslateblue", "gold"),
    alpha = .3,
    facet.by = "affect",
    panel.labs = list(affect = c("Negative Affect", "Positive Affect")),
    xlab = "Run",
    ylab = "Affect Rating",
    legend.title = "Run",
    title = ""
    ) +
  geom_point(aes(col = run_type), position = "jitter", size = .8, alpha = .8) 
  
ggsave("vas_rating.png", width = 5, height = 4)
  
```

#################### CORRS AMONG ROIS ##########################

```{r}
rois_pad_WIN <-
  rois_pad %>%
  filter(condition == "FB_WIN")

ROI_corrtable_WIN <-
  correltable(
    rois_pad_WIN,
    vars = c("NAcc_l.pre", "NAcc_l.post", "NAcc_r.pre", "NAcc_r.post", "caudate_l.pre", "caudate_l.post",  "caudate_r.pre", "caudate_r.post", "putamen_l.pre", "putamen_l.post", "putamen_r.pre", "putamen_r.post"),
    tri = "lower"
  )
ROI_corrtable_WIN

rois_pad_LOSS <-
  rois_pad %>%
  filter(condition == "FB_LOSS")

ROI_corrtable_LOSS <-
  correltable(
    rois_pad_LOSS,
    vars = c("AntIns_l.pre", "AntIns_l.post", "AntIns_r.pre", "AntIns_r.post", "dACC.pre", "dACC.post"),
    tri = "lower"
  )
ROI_corrtable_LOSS
```
