==> 3_Analysis_prepost_mafq.Rmd <==
---
title: "Analysis_prepost_mafq"
author: "Jackie Kirshenbaum"
date: "`r format(Sys.Date(), '%B %e %Y')`"
output:
  html_document:
    df_print: paged
  toc: yes
  toc_float: yes
---

# Libraries 

```{r, libraries, message=FALSE}
library(tidyverse) # 2.0.0 
library(sjPlot) # 2.8.12
library(interactions) # 1.1.5
library(performance) # 0.10.2
library(lmerTest) # 3.1-3
library(interactions) # 1.1.5

sessionInfo()
```

# Read in Data

```{r, filepaths, message=FALSE}
pad_and_rois_merge_data_fp <- "~/Jackie/pad_and_rois_usable_mafq.csv"
rois_change_LOSS_fp <- "~/Jackie/rois_change_LOSS.csv"
rois_change_WIN_fp <- "~/Jackie/rois_change_WIN.csv"

pad_and_rois_merge_data <-
  read_csv(pad_and_rois_merge_data_fp) %>%
  dplyr::mutate(
    masc_c_total = masqharmavoid_c + masqphysical_c + masqsepanx_c + masqsocial_c,
    subID = factor(subID),
    Group = factor(Group),
    condition = factor(condition)    
    )

pad_and_rois_merge_data_WIN <- 
  pad_and_rois_merge_data %>%
  filter(condition == "FB_WIN") %>%
  dplyr::select(-c(ends_with(".pre"), ends_with(".post")))

pad_and_rois_merge_data_LOSS <- 
  pad_and_rois_merge_data %>%
  filter(condition == "FB_LOSS") %>%
  dplyr::select(-c(ends_with(".pre"), ends_with(".post")))

roi_LOSS <-
  read_csv(rois_change_LOSS_fp) %>%
  dplyr::mutate(
    subID = factor(subID),
    condition = factor(condition)
    )  
roi_WIN <-
  read_csv(rois_change_WIN_fp) %>%
  dplyr::mutate(
    subID = factor(subID),
    condition = factor(condition)
    )  

pad_and_rois_merge_WIN <-
  left_join(
    pad_and_rois_merge_data_WIN,
    roi_WIN,
    by = c("subID", "condition")
  )

pad_and_rois_merge_LOSS <-
  left_join(
    pad_and_rois_merge_data_LOSS,
    roi_LOSS,
    by = c("subID", "condition")
  )
```

## z scoring variables
```{r}
pad_and_rois_merge_WIN_z <-
  pad_and_rois_merge_WIN %>%
  dplyr::mutate(
    Group = factor(Group),
    Sex = factor(Sex),
    Age_z = as.numeric(scale(Age)),
    TannerMean_z = as.numeric(scale(TannerMean)),
    aleq_c.L1_z = as.numeric(scale(aleq_c.L1)), # L1 total stress
    aleq_c.L2_z = as.numeric(scale(aleq_c.L2)), # L2 total stress
    aleqNonInt_c.L1_z = as.numeric(scale(aleqNonInt_c.L1)), # L1 non-interpersonal stress
    aleqNonInt_c.L2_z = as.numeric(scale(aleqNonInt_c.L2)), # L2 non-interpersonal stress 
    aleqInt_c.L1_z = as.numeric(scale(aleqInt_c.L1)), # L1 interpersonal stress
    aleqInt_c.L2_z = as.numeric(scale(aleqInt_c.L2)), # L2 interpersonal stress
    visit_num_z = as.numeric(scale(visitN)),
    mafq_c_w_z = as.numeric(scale(mafq_c_w)), # depression
    masc_c_total_z = as.numeric(scale(masc_c_total)), # anxiety total
    shaps_c_z = as.numeric(scale(shaps_c)) # anhedonia
    ) 

contrasts(pad_and_rois_merge_WIN_z$Group) = contr.treatment(2, base = 2)
contrasts(pad_and_rois_merge_WIN_z$Sex) = contr.treatment(2, base = 2)

pad_and_rois_merge_LOSS_z <-
  pad_and_rois_merge_LOSS %>%
  dplyr::mutate(
    Group = factor(Group),
    Sex = factor(Sex),
    Age_z = as.numeric(scale(Age)),
    TannerMean_z = as.numeric(scale(TannerMean)),
    aleq_c.L1_z = as.numeric(scale(aleq_c.L1)), # L1 total stress
    aleq_c.L2_z = as.numeric(scale(aleq_c.L2)), # L2 total stress
    aleqNonInt_c.L1_z = as.numeric(scale(aleqNonInt_c.L1)), # L1 non-interpersonal stress
    aleqNonInt_c.L2_z = as.numeric(scale(aleqNonInt_c.L2)), # L2 non-interpersonal stress    
    aleqInt_c.L1_z = as.numeric(scale(aleqInt_c.L1)), # L1 interpersonal stress
    aleqInt_c.L2_z = as.numeric(scale(aleqInt_c.L2)), # L2 interpersonal stress
    visit_num_z = as.numeric(scale(visitN)),
    mafq_c_w_z = as.numeric(scale(mafq_c_w)), # depression
    masc_c_total_z = as.numeric(scale(masc_c_total)), # anxiety
    shaps_c_z = as.numeric(scale(shaps_c))   # anhedonia
  )

contrasts(pad_and_rois_merge_LOSS_z$Group) = contr.treatment(2, base = 2)
contrasts(pad_and_rois_merge_LOSS_z$Sex) = contr.treatment(2, base = 2)
```

```{r}
set_theme(
  base = theme_bw()
)
options(scipen = 999)
```

# Does brain pre-post stress in the win condition correlate with depression severity 

####  between-person DF

```{r}
pad_and_rois_merge_WIN_z_l2 <-
  pad_and_rois_merge_WIN_z %>%
  dplyr::select(-ends_with(c(".L1", ".L1z"))) %>%
  dplyr::group_by(subID) %>%
  dplyr::mutate(
    mean_mafq = mean(mafq_c_w, na.rm = TRUE),
    mean_masc = mean(masc_c_total, na.rm = TRUE),
    mean_shaps = mean(shaps_c, na.rm = TRUE)
  ) %>%
  filter(visitN == 0) 
```

### NAcc pre-post-stress win
```{r}
# NAcc_l x within-person stress 
mlm_NAcc_l_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z*NAcc_l_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(mlm_NAcc_l_win_prepost_z) # no interaction
sjPlot::tab_model(
  mlm_NAcc_l_win_prepost_z, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234)

# NAcc_r x within-person stress
mlm_NAcc_r_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z*NAcc_r_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(mlm_NAcc_r_win_prepost_z) # no interaction
sjPlot::tab_model(
  mlm_NAcc_r_win_prepost_z, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234)
```

> No interaction between NAcc and stress predicting depression

### Caudate pre-post-stress win
> All participants had full coverage of caudate ROIs

```{r}
# caudate_l x within-person stress
mlm_caudate_l_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * caudate_l_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(mlm_caudate_l_win_prepost_z) # interactions: aleq_c.L1_z:caudate_l_prepost_resWIN 
sjPlot::tab_model(
  mlm_caudate_l_win_prepost_z, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234)
sim_slopes(mlm_caudate_l_win_prepost_z, pred=aleq_c.L1_z, modx=caudate_l_prepost_resWIN, jnplot=TRUE, confint = TRUE)
probe_interaction(mlm_caudate_l_win_prepost_z, aleq_c.L1_z, caudate_l_prepost_resWIN)

check_convergence(mlm_caudate_l_win_prepost_z)
plot_model(mlm_caudate_l_win_prepost_z, type = "diag") 
check_model(mlm_caudate_l_win_prepost_z) # 3 outliers
check_outliers(mlm_caudate_l_win_prepost_z)
pad_and_rois_merge_WIN_z[-c(61, 118, 332),]

# Testing without outliers
mlm_caudate_l_win_prepost_z_out <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * caudate_l_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )
check_model(mlm_caudate_l_win_prepost_z_out) # looks like no other outliers
check_convergence(mlm_caudate_l_win_prepost_z_out)
sjPlot::tab_model(
  mlm_caudate_l_win_prepost_z_out, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234)


####### Exploratory adjusting for longitudinal MASC and SHAPS ####### 
mlm_caudate_l_win_prepost_z_masc <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * caudate_l_prepost_resWIN  + masc_c_total_z +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )
summary(mlm_caudate_l_win_prepost_z_masc)  
check_convergence(mlm_caudate_l_win_prepost_z_masc)
sjPlot::tab_model(
  mlm_caudate_l_win_prepost_z_masc, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234) # interaction holds 

mlm_caudate_l_win_prepost_masc <- 
  lmer(
    mafq_c_w ~ 
      Age_z + TannerMean + aleq_c.L2 +  aleq_c.L1 * caudate_l_prepost_resWIN  + masc_c_total +
         (1 + visitN | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )

# title = "L Caudate Activation Post-Stress Moderates \n Longitudinal Association between Life Stress and Depression"
sjPlot::plot_model(mlm_caudate_l_win_prepost_masc, type = "int", show.data = TRUE, colors = "Dark2", dot.size = 1, robust = TRUE, mdrt.values = "meansd", title = "", legend.title ="Residualized L Caudate \n Post-Pre Acute Stress")
save_plot("L.CaudxStress-Depression.png", fig = ggplot2::last_plot(), width = 12, height = 8, dpi = 300, theme = ggplot2::theme_get(), label.color = "black", label.size = 2.4, axis.textsize = 0.8, axis.titlesize = 0.75, legend.textsize = 0.6, legend.titlesize = 0.65, legend.itemsize = 0.5)

sim_slopes(mlm_caudate_l_win_prepost_z_masc, pred=aleq_c.L1_z, modx=caudate_l_prepost_resWIN, jnplot=TRUE)

mlm_caudate_l_win_prepost_z_shaps <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * caudate_l_prepost_resWIN  + shaps_c_z +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )
summary(mlm_caudate_l_win_prepost_z_shaps) 
check_convergence(mlm_caudate_l_win_prepost_z_shaps)
sjPlot::tab_model( 
  mlm_caudate_l_win_prepost_z_shaps, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234) # interaction holds
sim_slopes(mlm_caudate_l_win_prepost_z_shaps, pred=aleq_c.L1_z, modx=caudate_l_prepost_resWIN, jnplot=TRUE)

# caudate_r x within-person stress
mlm_caudate_r_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * caudate_r_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(mlm_caudate_r_win_prepost_z)# interactions: aleq_c.L1_z:caudate_r_prepost_resWIN 
sim_slopes(mlm_caudate_r_win_prepost_z, pred=aleq_c.L1_z, modx=caudate_r_prepost_resWIN, jnplot=TRUE, confint = TRUE)
probe_interaction(mlm_caudate_r_win_prepost_z, aleq_c.L1_z, caudate_r_prepost_resWIN)
sjPlot::tab_model(
  mlm_caudate_r_win_prepost_z, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234)
check_convergence(mlm_caudate_r_win_prepost_z)
plot_model(mlm_caudate_r_win_prepost_z, type = "diag") 
check_model(mlm_caudate_r_win_prepost_z) # 3 outliers
check_outliers(mlm_caudate_r_win_prepost_z)
pad_and_rois_merge_WIN_z[-c(61, 118, 332),]

# Testing without outliers
mlm_caudate_r_win_prepost_z_out <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * caudate_r_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )
check_model(mlm_caudate_r_win_prepost_z_out) # looks like no other outliers
check_convergence(mlm_caudate_r_win_prepost_z_out)
sjPlot::tab_model(
  mlm_caudate_r_win_prepost_z_out, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234)

####### Exploratory adjusting for longitudinal MASC and SHAPS ####### 
mlm_caudate_r_win_prepost_z_masc <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * caudate_r_prepost_resWIN  + masc_c_total_z +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )
summary(mlm_caudate_r_win_prepost_z_masc) # interaction holds 
check_convergence(mlm_caudate_r_win_prepost_z_masc)
sjPlot::tab_model(
  mlm_caudate_r_win_prepost_z_masc, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234)

mlm_caudate_r_win_prepost_masc <- 
  lmer(
    mafq_c_w ~ 
      Age_z + TannerMean + aleq_c.L2 +  aleq_c.L1 * caudate_r_prepost_resWIN  + masc_c_total +
         (1 + visitN | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )

# title = "R Caudate Activation Post-Stress Moderates \n Longitudinal Association between Life Stress and Depression"
sjPlot::plot_model(mlm_caudate_r_win_prepost_masc, type = "int", show.data = TRUE, colors = "Dark2", dot.size = 1, robust = TRUE, mdrt.values = "meansd", title = "", legend.title ="Residualized R Caudate \n Post-Pre Acute Stress")
save_plot("R.CaudxStress-Depression.png", fig = ggplot2::last_plot(), width = 12, height = 8, dpi = 300, theme = ggplot2::theme_get(), label.color = "black", label.size = 2.4, axis.textsize = 0.8, axis.titlesize = 0.75, legend.textsize = 0.6, legend.titlesize = 0.65, legend.itemsize = 0.5)

sim_slopes(mlm_caudate_r_win_prepost_z_masc, pred=aleq_c.L1_z, modx=caudate_r_prepost_resWIN, jnplot=TRUE)

mlm_caudate_r_win_prepost_z_shaps <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * caudate_r_prepost_resWIN  + shaps_c_z +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )
summary(mlm_caudate_r_win_prepost_z_shaps) 
check_convergence(mlm_caudate_r_win_prepost_z_shaps)
sjPlot::tab_model( 
  mlm_caudate_r_win_prepost_z_shaps, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234) # interaction holds
sim_slopes(mlm_caudate_r_win_prepost_z_shaps, pred=aleq_c.L1_z, modx=caudate_r_prepost_resWIN, jnplot=TRUE)
```

> 2 way interaction between L and R caudate and stress predicitng depression

# Table: L Caudate
```{r}
sjPlot::tab_model(
  mlm_caudate_l_win_prepost_z, 
  mlm_caudate_l_win_prepost_z_out, 
  mlm_caudate_l_win_prepost_z_masc, 
  mlm_caudate_l_win_prepost_z_shaps,
  bootstrap = TRUE,
  iterations = 1000, 
  dv.labels = c("Depression Severity", "Depression Severity", "Depression Severity", "Depression Severity"), 
  seed = 1234,
  pred.labels =
    c("Intercept", 
      "Age", 
      "Tanner",
      "Between-person Stress", 
      "Within-person Stress", 
      "L Caudate", 
      "Within-person Stress*L Caudate", 
      "Anxiety", 
      "Anhedonia"
      ),
  file = "LCaudate_Main_Depression.doc")
```

# Table: R Caudate
```{r}
sjPlot::tab_model(
  mlm_caudate_r_win_prepost_z, 
  mlm_caudate_r_win_prepost_z_out, 
  mlm_caudate_r_win_prepost_z_masc, 
  mlm_caudate_r_win_prepost_z_shaps,
  bootstrap = TRUE,
  iterations = 1000, 
  dv.labels = c("Depression Severity", "Depression Severity", "Depression Severity", "Depression Severity"), 
  seed = 1234,
  pred.labels =
    c("Intercept", 
      "Age", 
      "Tanner",
      "Between-person Stress", 
      "Within-person Stress", 
      "R Caudate", 
      "Within-person Stress*R Caudate", 
      "Anxiety", 
      "Anhedonia"
      ),
  file = "RCaudate_Main_Depression.doc")
```

> 2 way interaction between L and R caudate and stress predicitng depression

#### Based on Reviewer Recommendation, testing interpersonal vs noninterersonal stress

```{r}
# caudate_l x within-person interpersonal stress
INT_mlm_caudate_l_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleqInt_c.L2_z +  aleqInt_c.L1_z * caudate_l_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(INT_mlm_caudate_l_win_prepost_z) # interaction sig

# caudate_l x within-person non-interpersonal stress
NONINT_mlm_caudate_l_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleqNonInt_c.L2_z +  aleqNonInt_c.L1_z * caudate_l_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(NONINT_mlm_caudate_l_win_prepost_z) # interaction not sig

# caudate_r x within-person interpersonal stress
INTmlm_caudate_r_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleqInt_c.L2_z +  aleqInt_c.L1_z * caudate_r_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(INTmlm_caudate_r_win_prepost_z) # interaction sig

# caudate_r x within-person non-interpersonal stress
NONINTmlm_caudate_r_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleqNonInt_c.L2_z +  aleqNonInt_c.L1_z * caudate_r_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(NONINTmlm_caudate_r_win_prepost_z) # interaction not sig
```

> 2 way interaction between L and R caudate and interpersonal stress, but not non-interpersonal stress predicitng depression

#### Based on Reviewer Recommendation, testing model with fixed effect of visit 

```{r}
visit_mlm_caudate_l_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * caudate_l_prepost_resWIN  +
         visit_num_z + (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(visit_mlm_caudate_l_win_prepost_z)

visit_mlm_caudate_r_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * caudate_r_prepost_resWIN  +
         visit_num_z + (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(visit_mlm_caudate_r_win_prepost_z)
```

> Adding fixed effect of visit number does not change the interaction between stress and ROI

### Putamen pre-post-stress win
> One participant was missing 2 voxels of coverage in their R putamen, so this code chucnk also conducts sensitivity analysis excluding that participant.

```{r}
# putamen_l x within-person stress
mlm_putamen_l_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z*putamen_l_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(mlm_putamen_l_win_prepost_z) # interactions: aleq_c.L1_z:putamen_l_prepost_resWIN
sjPlot::tab_model(
  mlm_putamen_l_win_prepost_z,
  bootstrap = TRUE,
  iterations = 1000,
  seed = 1234
)
sim_slopes(mlm_putamen_l_win_prepost_z, pred=aleq_c.L1_z, modx=putamen_l_prepost_resWIN, jnplot=TRUE, confint = TRUE)
probe_interaction(mlm_putamen_l_win_prepost_z, aleq_c.L1_z, putamen_l_prepost_resWIN)
check_convergence(mlm_putamen_l_win_prepost_z)
plot_model(mlm_putamen_l_win_prepost_z, type = "diag") 
check_model(mlm_putamen_l_win_prepost_z) # 3 outliers (same as above)
check_outliers(mlm_putamen_l_win_prepost_z)

mlm_putamen_l_win_prepost_z_out <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z*putamen_l_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )
check_model(mlm_putamen_l_win_prepost_z_out) # looks like no other outliers
check_convergence(mlm_putamen_l_win_prepost_z_out)
sjPlot::tab_model(
  mlm_putamen_l_win_prepost_z_out, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234)

####### Exploratory adjusting for longitudinal MASC and SHAPS ####### 
mlm_putamen_l_win_prepost_z_masc <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * putamen_l_prepost_resWIN  + masc_c_total_z +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )
summary(mlm_putamen_l_win_prepost_z_masc) 
check_convergence(mlm_putamen_l_win_prepost_z_masc)
sjPlot::tab_model(
  mlm_putamen_l_win_prepost_z_masc,
  bootstrap = TRUE,
  iterations = 1000,
  seed = 1234 # interaction holds 
)

mlm_putamen_l_win_prepost_masc <- 
  lmer(
    mafq_c_w ~ 
      Age + TannerMean  + aleq_c.L2 +  aleq_c.L1 * putamen_l_prepost_resWIN  + masc_c_total +
         (1 + visitN | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )
sjPlot::plot_model(mlm_putamen_l_win_prepost_masc, type = "int", show.data = TRUE, colors = "Dark2", dot.size = 1,robust = TRUE, mdrt.values = "meansd", title = "L Putamen Activation Post-Stress Moderates \n Longitudinal Association between Life Stress and Depression", legend.title ="Residualized L Putamen \n Post-Pre Acute Stress")

sim_slopes(mlm_putamen_l_win_prepost_z_masc, pred=aleq_c.L1_z, modx=putamen_l_prepost_resWIN, jnplot=TRUE)

mlm_putamen_l_win_prepost_z_shaps <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * putamen_l_prepost_resWIN  + shaps_c_z +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(mlm_putamen_l_win_prepost_z_shaps) 
check_convergence(mlm_putamen_l_win_prepost_z_shaps)
sjPlot::tab_model(
  mlm_putamen_l_win_prepost_z_shaps,
  bootstrap = TRUE,
  iterations = 1000,
  seed = 1234 # interaction holds 
)
mlm_putamen_l_win_prepost_shaps <- 
  lmer(
    mafq_c_w ~ 
      Age + TannerMean  + aleq_c.L2 +  aleq_c.L1 * putamen_l_prepost_resWIN  + shaps_c +
         (1 + visitN | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )
sjPlot::plot_model(mlm_putamen_l_win_prepost_shaps, type = "int", show.data = TRUE, colors = "Dark2", dot.size = 1, robust = TRUE, mdrt.values = "meansd", title = "L Putamen Activation Post-Stress Moderates \n Longitudinal Association between Life Stress and Depression", legend.title ="Residualized L Putamen \n Post-Pre Acute Stress")

sim_slopes(mlm_putamen_l_win_prepost_z_shaps, pred=aleq_c.L1_z, modx=putamen_l_prepost_resWIN, jnplot=TRUE)


# putamen_r x within-person stress 
mlm_putamen_r_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z*putamen_r_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(mlm_putamen_r_win_prepost_z) # interactions:aleq_c.L1_z:putamen_r_prepost_resWIN   
sjPlot::tab_model(
  mlm_putamen_r_win_prepost_z,
  bootstrap = TRUE,
  iterations = 1000,
  seed = 1234
)
sim_slopes(mlm_putamen_r_win_prepost_z, pred=aleq_c.L1_z, modx=putamen_r_prepost_resWIN, jnplot=TRUE, confint = TRUE)
probe_interaction(mlm_putamen_r_win_prepost_z, aleq_c.L1_z, putamen_r_prepost_resWIN)
check_convergence(mlm_putamen_r_win_prepost_z)
plot_model(mlm_putamen_r_win_prepost_z, type = "diag") 
check_model(mlm_putamen_r_win_prepost_z) # 3 outliers (same as above)
check_outliers(mlm_putamen_r_win_prepost_z)

mlm_putamen_r_win_prepost_z_out <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z*putamen_r_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )
check_model(mlm_putamen_r_win_prepost_z_out) # looks like no other outliers
check_convergence(mlm_putamen_r_win_prepost_z_out)
sjPlot::tab_model(
  mlm_putamen_r_win_prepost_z_out, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234)

####### Exploratory adjusting for longitudinal MASC and SHAPS ####### 
mlm_putamen_r_win_prepost_z_masc <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * putamen_r_prepost_resWIN  + masc_c_total_z +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )
summary(mlm_putamen_r_win_prepost_z_masc) 
check_convergence(mlm_putamen_r_win_prepost_z_masc)
sjPlot::tab_model(
  mlm_putamen_r_win_prepost_z_masc,
  bootstrap = TRUE,
  iterations = 1000,
  seed = 1234 # interaction holds 
)

mlm_putamen_r_win_prepost_masc <- 
  lmer(
    mafq_c_w ~ 
      Age + TannerMean  + aleq_c.L2 +  aleq_c.L1 * putamen_r_prepost_resWIN  + masc_c_total +
         (1 + visitN | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )
sjPlot::plot_model(mlm_putamen_r_win_prepost_masc, type = "int", show.data = TRUE, colors = "Dark2", dot.size = 1,robust = TRUE, mdrt.values = "meansd", title = "R Putamen Activation Post-Stress Moderates \n Longitudinal Association between Life Stress and Depression", legend.title ="Residualized R Putamen \n Post-Pre Acute Stress")

sim_slopes(mlm_putamen_r_win_prepost_z_masc, pred=aleq_c.L1_z, modx=putamen_r_prepost_resWIN, jnplot=TRUE)

mlm_putamen_r_win_prepost_z_shaps <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * putamen_r_prepost_resWIN  + shaps_c_z +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )
summary(mlm_putamen_r_win_prepost_z_shaps) 
check_convergence(mlm_putamen_r_win_prepost_z_shaps)
sjPlot::tab_model(
  mlm_putamen_r_win_prepost_z_shaps,
  bootstrap = TRUE,
  iterations = 1000,
  seed = 1234 # interaction holds 
)
mlm_putamen_r_win_prepost_shaps <- 
  lmer(
    mafq_c_w ~ 
      Age + TannerMean  + aleq_c.L2 +  aleq_c.L1 * putamen_r_prepost_resWIN  + shaps_c +
         (1 + visitN | subID), 
       data = pad_and_rois_merge_WIN_z[-c(61, 118, 332),]
  )
sjPlot::plot_model(mlm_putamen_r_win_prepost_shaps, type = "int", show.data = TRUE, colors = "Dark2", dot.size = 1, robust = TRUE, mdrt.values = "meansd", title = "R Putamen Activation Post-Stress Moderates \n Longitudinal Association between Life Stress and Depression", legend.title ="Residualized R Putamen \n Post-Pre Acute Stress")

sim_slopes(mlm_putamen_r_win_prepost_z_shaps, pred=aleq_c.L1_z, modx=putamen_r_prepost_resWIN, jnplot=TRUE)



#### Sensitivity Excluding HP809 ####
# sub HP809 includes rows
pad_and_rois_merge_WIN_Z_sens <- pad_and_rois_merge_WIN_z[pad_and_rois_merge_WIN_z$subID != "sub-HP809", ]

mlm_putamen_r_win_prepost_z_sens <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z*putamen_r_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_Z_sens
  )
summary(mlm_putamen_r_win_prepost_z_sens) 
check_outliers(mlm_putamen_r_win_prepost_z_sens) # (cases 61, 114, 328).

mlm_putamen_r_win_prepost_z_sens <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z*putamen_r_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_Z_sens[-c(61, 114, 328),]
  )
summary(mlm_putamen_r_win_prepost_z_sens) 
sjPlot::tab_model(
  mlm_putamen_r_win_prepost_z_sens,
  bootstrap = TRUE,
  iterations = 1000,
  seed = 1234 # interaction holds 
)
mlm_putamen_r_win_prepost_z_masc_sens <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * putamen_r_prepost_resWIN  + masc_c_total_z +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_Z_sens[-c(61, 114, 328),]
  )
summary(mlm_putamen_r_win_prepost_z_masc_sens)
sjPlot::tab_model(
  mlm_putamen_r_win_prepost_z_masc_sens,
  bootstrap = TRUE,
  iterations = 1000,
  seed = 1234 # interaction holds 
)

mlm_putamen_r_win_prepost_z_shaps_sens <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * putamen_r_prepost_resWIN  + shaps_c_z +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_Z_sens[-c(61, 114, 328),]
  )
summary(mlm_putamen_r_win_prepost_z_shaps_sens) 
sjPlot::tab_model(
  mlm_putamen_r_win_prepost_z_shaps_sens,
  bootstrap = TRUE,
  iterations = 1000,
  seed = 1234 # interaction holds 
)
```

> 2 way interaction between L and R putamen and stress predicitng depressoin

# Table: L Putamen
```{r}
sjPlot::tab_model(
  mlm_putamen_l_win_prepost_z, 
  mlm_putamen_l_win_prepost_z_out, 
  mlm_putamen_l_win_prepost_z_masc, 
  mlm_putamen_l_win_prepost_z_shaps,
  bootstrap = TRUE,
  iterations = 1000, 
  dv.labels = c("Depression Severity", "Depression Severity", "Depression Severity", "Depression Severity"), 
  seed = 1234,
  pred.labels =
    c("Intercept", 
      "Age", 
      "Tanner",
      "Between-person Stress", 
      "Within-person Stress", 
      "L Putamen", 
      "Within-person Stress*L Putamen", 
      "Anxiety", 
      "Anhedonia"
      ),
  file = "LPutamen_Main_Depression.doc")
```

# Table: R Putamen
```{r}
sjPlot::tab_model(
  mlm_putamen_r_win_prepost_z, 
  mlm_putamen_r_win_prepost_z_out, 
  mlm_putamen_r_win_prepost_z_masc, 
  mlm_putamen_r_win_prepost_z_shaps,
  bootstrap = TRUE,
  iterations = 1000, 
  dv.labels = c("Depression Severity", "Depression Severity", "Depression Severity", "Depression Severity"), 
  seed = 1234,
  pred.labels =
    c("Intercept", 
      "Age", 
      "Tanner",
      "Between-person Stress", 
      "Within-person Stress", 
      "R Putamen", 
      "Within-person Stress*R Putamen", 
      "Anxiety", 
      "Anhedonia"
      ),
  file = "RPutamen_Main_Depression.doc")
```


# Table: Removing Observations without full Coverage in R Putamen
```{r}
sjPlot::tab_model(
  mlm_putamen_r_win_prepost_z_sens, 
  mlm_putamen_r_win_prepost_z_masc_sens, 
  mlm_putamen_r_win_prepost_z_shaps_sens, 
  bootstrap = TRUE,
  iterations = 1000, 
  dv.labels = c("Depression Severity", "Depression Severity", "Depression Severity"), 
  seed = 1234,
  pred.labels =
    c("Intercept", 
      "Age", 
      "Tanner",
      "Between-person Stress", 
      "Within-person Stress", 
      "R Putamen", 
      "Within-person Stress*R Putamen", 
      "Anxiety", 
      "Anhedonia"
      ),
  file = "RPutamen_dep_control_anx_anhedonia_sensitivity.doc")
```

#### Based on Reviewer Recommendation, testing interpersonal vs noninterersonal stress

```{r}
# putamen_l x within-person interpersonal stress
INT_mlm_putamen_l_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleqInt_c.L2_z +  aleqInt_c.L1_z * putamen_l_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(INT_mlm_putamen_l_win_prepost_z) # interaction sig

# putamen_l x within-person non-interpersonal stress
NONINT_mlm_putamen_l_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleqNonInt_c.L2_z +  aleqNonInt_c.L1_z * putamen_l_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(NONINT_mlm_putamen_l_win_prepost_z) # interaction not sig

# putamen_r x within-person interpersonal stress
INTmlm_putamen_r_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleqInt_c.L2_z +  aleqInt_c.L1_z * putamen_r_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(INTmlm_putamen_r_win_prepost_z) # interaction sig

# putamen_r x within-person non-interpersonal stress
NONINTmlm_putamen_r_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleqNonInt_c.L2_z +  aleqNonInt_c.L1_z * putamen_r_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(NONINTmlm_putamen_r_win_prepost_z) # interaction not sig
```

#### Based on Reviewer Recommendation, testing model with fixed effect of visit 

```{r}
visit_mlm_putamen_l_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * putamen_l_prepost_resWIN  +
         visit_num_z + (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(visit_mlm_putamen_l_win_prepost_z)

visit_mlm_putamen_r_win_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z * putamen_r_prepost_resWIN  +
         visit_num_z + (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(visit_mlm_putamen_r_win_prepost_z)
```

> Adding fixed effect of visit number does not change the interaction between stress and ROI


#### MC CORRECTION FOR WIN

```{r}
# l NAcc, r NAcc, l Caudate, r Caudate, l Putamen, r Putamen
pgain <- c(0.940, 0.314, 0.001, 0.002, 0.001, 0.001)
p.adjust(pgain, method = "fdr", n = length(pgain))
# CORRECTED l NACC: 0.9400 r NACC: 0.3768 l CAUD: 0.0020, r CAUD: 0.0030, l PUT: 0.0020, r PUT: 0.0020
```


# Does brain pre-post stress in the loss condition correlate with depression severity 

#### For between-person assoications

```{r}
pad_and_rois_merge_LOSS_z_l2 <-
  pad_and_rois_merge_LOSS_z %>%
  dplyr::select(-ends_with(c(".L1", ".L1z"))) %>%
  dplyr::group_by(subID) %>%
  dplyr::mutate(
    mean_mafq = mean(mafq_c_w, na.rm = TRUE),
    mean_mafq_log = mean(mafq_c_w_log, na.rm = TRUE)
  ) %>%
  filter(visitN == 0) 
```

### dACC pre-post-stress loss
```{r}
# dACC x within-person stress
mlm_dacc_loss_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z*dACC_prepost_resLOSS  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
summary(mlm_dacc_loss_prepost_z) # No interaction 
```

### AntIns pre-post-stress loss
```{r}
mlm_antins_l_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z*AntIns_l_prepost_resLOSS  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
summary(mlm_antins_l_prepost_z) #  No interaction   

mlm_antins_r_prepost_z <- 
  lmer(
    mafq_c_w_z ~ 
      Age_z + TannerMean_z + aleq_c.L2_z +  aleq_c.L1_z*AntIns_r_prepost_resLOSS  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
summary(mlm_antins_r_prepost_z) #  No interaction

```

# Null Win Model Tables
```{r}
sjPlot::tab_model(
  mlm_NAcc_l_win_prepost_z, 
  mlm_NAcc_r_win_prepost_z,
  bootstrap = TRUE,
  iterations = 1000, 
  dv.labels = c("Depression Severity", "Depression Severity"), 
  seed = 1234,
  pred.labels = 
    c("Intercept",
      "Age",
      "Tanner",
      "Between-person Stress", 
      "Within-person Stress",
      "L NAcc",
      "Within-Person Stress* L NAcc",
      "R NAcc",
      "Within-Person Stress* R NAcc"
      ),
  file = "Depression_NullResults_Win.doc")
```

# Null Loss Model Tables
```{r}
sjPlot::tab_model(
  mlm_dacc_loss_prepost_z, 
  mlm_antins_l_prepost_z,
  mlm_antins_r_prepost_z,
  bootstrap = TRUE,
  iterations = 1000, 
  dv.labels = c("Depression Severity", "Depression Severity", "Depression Severity"), 
  seed = 1234,
  pred.labels = 
    c("Intercept",
      "Age",
      "Tanner",
      "Between-person Stress", 
      "Within-person Stress",
      "dACC",
      "Within-Person Stress* dACC",
      "L Anterior Insula",
      "Within-Person Stress* L Anterior Insula",
      "R Anterior Insula",
      "Within-Person Stress* R Anterior Insula"
      ),
  file = "Depression_NullResults_Loss.doc")
```
