==> 3_Analysis_prepost_masc.Rmd <==
---
title: "Analysis_prepost_masc"
author: "Jackie Kirshenbaum"
date: "`r format(Sys.Date(), '%B %e %Y')`"
output:
  html_document:
    df_print: paged
  toc: yes
  toc_float: yes
---

# Libraries 

```{r, libraries, message=FALSE}
library(tidyverse) # 2.0.0 
library(sjPlot) # 2.8.12
library(interactions) # 1.1.5
library(performance) # 0.10.2
library(lmerTest) # 3.1-3
library(interactions) # 1.1.5

sessionInfo()
```

# Read in Data

```{r, filepaths, message=FALSE}
pad_and_rois_merge_data_fp <- "~/Jackie/pad_and_rois_usable_masc.csv"
rois_change_LOSS_fp <- "~/Jackie/rois_change_LOSS.csv"
rois_change_WIN_fp <- "~/Jackie/rois_change_WIN.csv"
mfq_fp <- "~/Jackie/pad_and_rois_usable_mafq.csv"

mfq <-
  read_csv(mfq_fp) %>%
  dplyr::select(subID, visit, visitN, mafq_c_w, mafq_c_w_log, condition) %>%
  dplyr::mutate(
    subID = factor(subID),
    visit = factor(visit),
    condition = factor(condition)
  )

pad_and_rois_merge_data <- read_csv(pad_and_rois_merge_data_fp) %>%
    dplyr::mutate(
    subID = factor(subID),
    visit = factor(visit),
    condition = factor(condition)
  )
pad_and_rois_merge_data <-
  left_join(pad_and_rois_merge_data, mfq, by = c("subID", "visit", "visitN", "condition"))
  
pad_and_rois_merge_data_WIN <- 
  pad_and_rois_merge_data %>%
  dplyr::mutate(
    subID = factor(subID),
    Group = factor(Group),
    condition = factor(condition)
    ) %>%
  filter(condition == "FB_WIN") %>%
  dplyr::select(-c(ends_with(".pre"), ends_with(".post")))

pad_and_rois_merge_data_LOSS <- 
  pad_and_rois_merge_data %>%
  dplyr::mutate(
    subID = factor(subID),
    Group = factor(Group),
    condition = factor(condition)
    ) %>%
  filter(condition == "FB_LOSS") %>%
  dplyr::select(-c(ends_with(".pre"), ends_with(".post")))

roi_LOSS <-
  read_csv(rois_change_LOSS_fp) %>%
  dplyr::mutate(
    subID = factor(subID),
    condition = factor(condition)
    )  
roi_WIN <-
  read_csv(rois_change_WIN_fp) %>%
  dplyr::mutate(
    subID = factor(subID),
    condition = factor(condition)
    )  

pad_and_rois_merge_WIN <-
  left_join(
    pad_and_rois_merge_data_WIN,
    roi_WIN,
    by = c("subID", "condition")
  )

pad_and_rois_merge_LOSS <-
  left_join(
    pad_and_rois_merge_data_LOSS,
    roi_LOSS,
    by = c("subID", "condition")
   )

```

## z scoring variables
```{r}
pad_and_rois_merge_WIN_z <-
  pad_and_rois_merge_WIN %>%
  dplyr::mutate(
    Group = factor(Group),
    Sex = factor(Sex),
    Age_z = as.numeric(scale(Age)),
    TannerMean_z = as.numeric(scale(TannerMean)),
    aleq_c.L1_z = as.numeric(scale(aleq_c.L1)), # L1 total stress
    aleq_c.L2_z = as.numeric(scale(aleq_c.L2)), # L2 total stress
    aleqNonInt_c.L1_z = as.numeric(scale(aleqNonInt_c.L1)), # L1 non-interpersonal stress
    aleqNonInt_c.L2_z = as.numeric(scale(aleqNonInt_c.L2)), # L2 non-interpersonal stress 
    aleqInt_c.L1_z = as.numeric(scale(aleqInt_c.L1)), # L1 interpersonal stress
    aleqInt_c.L2_z = as.numeric(scale(aleqInt_c.L2)), # L2 interpersonal stress
    visit_num_z = as.numeric(scale(visitN)),
    masc_c_total_z = as.numeric(scale(masc_c_total)), # anxiety total
    masc_c_social_z = as.numeric(scale(masqsocial_c)), # social anxiety
    mafq_c_w_z = as.numeric(scale(mafq_c_w)), # depression
    shaps_c_z = as.numeric(scale(shaps_c)) # anhedonia
    ) 

contrasts(pad_and_rois_merge_WIN_z$Group) = contr.treatment(2, base = 2)
contrasts(pad_and_rois_merge_WIN_z$Sex) = contr.treatment(2, base = 2)

pad_and_rois_merge_LOSS_z <-
  pad_and_rois_merge_LOSS %>%
  dplyr::mutate(
    Group = factor(Group),
    Sex = factor(Sex),
    Age_z = as.numeric(scale(Age)),
    TannerMean_z = as.numeric(scale(TannerMean)),
    aleq_c.L1_z = as.numeric(scale(aleq_c.L1)), # L1 total stress
    aleq_c.L2_z = as.numeric(scale(aleq_c.L2)), # L2 total stress
    aleqNonInt_c.L1_z = as.numeric(scale(aleqNonInt_c.L1)), # L1 non-interpersonal stress
    aleqNonInt_c.L2_z = as.numeric(scale(aleqNonInt_c.L2)), # L2 non-interpersonal stress 
    aleqInt_c.L1_z = as.numeric(scale(aleqInt_c.L1)), # L1 interpersonal stress
    aleqInt_c.L2_z = as.numeric(scale(aleqInt_c.L2)), # L2 interpersonal stress
    visit_num_z = as.numeric(scale(visitN)),
    masc_c_total_z = as.numeric(scale(masc_c_total)), # anxiety total
    masc_c_social_z = as.numeric(scale(masqsocial_c)), # social anxiety
    mafq_c_w_z = as.numeric(scale(mafq_c_w)), # depression
    shaps_c_z = as.numeric(scale(shaps_c)) # anhedonia
  )

contrasts(pad_and_rois_merge_LOSS_z$Group) = contr.treatment(2, base = 2)
contrasts(pad_and_rois_merge_LOSS_z$Sex) = contr.treatment(2, base = 2)
```

```{r}
set_theme(
  base = theme_bw()
)
options(scipen = 999)
```

# Does brain pre-post stress in the win condition correlate with anxiety severity 

####  between-person DF

```{r}
pad_and_rois_merge_WIN_z_l2 <-
  pad_and_rois_merge_WIN_z %>%
  dplyr::select(-ends_with(c(".L1", ".L1z"))) %>%
  dplyr::group_by(subID) %>%
  dplyr::mutate(
    mean_masc = mean(masc_c_total, na.rm = TRUE)
  ) %>%
  filter(visitN == 0) 

```

### NAcc pre-post-stress win
```{r}
# NAcc_l x within-person stress
mlm_NAcc_l_win_prepost_z <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleq_c.L2_z +  aleq_c.L1_z*NAcc_l_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(mlm_NAcc_l_win_prepost_z) 
sjPlot::tab_model(
  mlm_NAcc_l_win_prepost_z, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234) # CI overlaps zero

# NAcc_r x within-person stress
mlm_NAcc_r_win_prepost_z <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleq_c.L2_z +  aleq_c.L1_z*NAcc_r_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(mlm_NAcc_r_win_prepost_z) 
sjPlot::tab_model(
  mlm_NAcc_r_win_prepost_z, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234) # No interaction

```
> No interaction by stress for NAcc


### Caudate pre-post-stress win
```{r}
# caudate_l x within-person stress 
mlm_caudate_l_win_prepost_z <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleq_c.L2_z +  aleq_c.L1_z*caudate_l_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(mlm_caudate_l_win_prepost_z) 
sjPlot::tab_model(
  mlm_caudate_l_win_prepost_z, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234) # no interaction 

# caudate_r x within-person stress
mlm_caudate_r_win_prepost_z <- 
  lmer(
    masc_c_total_z ~ 
       Sex + aleq_c.L2_z +  aleq_c.L1_z*caudate_r_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(mlm_caudate_r_win_prepost_z)
sjPlot::tab_model(
  mlm_caudate_r_win_prepost_z, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234) # no interaction 
```

> No interaction by stress for caudate

### Putamen pre-post-stress win
```{r}
# putamen_l x within-person stress
mlm_putamen_l_win_prepost_z <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleq_c.L2_z +  aleq_c.L1_z*putamen_l_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(mlm_putamen_l_win_prepost_z) 
sjPlot::tab_model(
  mlm_putamen_l_win_prepost_z, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234) # no interaction 

# putamen_r x within-person stress
mlm_putamen_r_win_prepost_z <- 
  lmer(
    masc_c_total_z ~ 
      Sex  + aleq_c.L2_z +  aleq_c.L1_z*putamen_r_prepost_resWIN  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_WIN_z
  )
summary(mlm_putamen_r_win_prepost_z) 
sjPlot::tab_model(
  mlm_putamen_r_win_prepost_z, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234) # no interaction 

```

> No interaction by stress for caudate

# Does brain pre-post stress in the loss condition correlate with anxiety severity 

#### For between-person assoications

```{r}
pad_and_rois_merge_LOSS_z_l2 <-
  pad_and_rois_merge_LOSS_z %>%
  dplyr::select(-ends_with(c(".L1", ".L1z"))) %>%
  dplyr::group_by(subID) %>%
  dplyr::mutate(
    mean_masc = mean(masc_c_total, na.rm = TRUE)
  ) %>%
  filter(visitN == 0) 
```

### dACC pre-post-stress loss
> No missing voxels in the dACC

```{r}
# dACC x within-person stress
mlm_dacc_loss_prepost_z <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleq_c.L2_z +  aleq_c.L1_z * dACC_prepost_resLOSS  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
summary(mlm_dacc_loss_prepost_z) # aleq_c.L1_z:dACC_prepost_resLOSS 
sjPlot::tab_model(
  mlm_dacc_loss_prepost_z, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234) 

sim_slopes(mlm_dacc_loss_prepost_z, pred=aleq_c.L1_z, modx=dACC_prepost_resLOSS, jnplot=TRUE, confint = TRUE)
probe_interaction(mlm_dacc_loss_prepost_z, aleq_c.L1_z, dACC_prepost_resLOSS)

check_convergence(mlm_dacc_loss_prepost_z)
plot_model(mlm_dacc_loss_prepost_z, type = "diag") 
check_model(mlm_dacc_loss_prepost_z) 


####### Exploratory adjusting for longitudinal MFQ and SHAPS ####### 
mlm_dacc_loss_prepost_z_mfq <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleq_c.L2_z +  aleq_c.L1_z * dACC_prepost_resLOSS  + mafq_c_w_z +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
summary(mlm_dacc_loss_prepost_z_mfq) # interaction holds 
check_convergence(mlm_dacc_loss_prepost_z_mfq)
sjPlot::tab_model(
  mlm_dacc_loss_prepost_z_mfq, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234) # interaction holds 

mlm_dacc_loss_prepost_mfq <- 
  lmer(
    masc_c_total ~ 
      Sex + aleq_c.L2 +  aleq_c.L1 * dACC_prepost_resLOSS  + mafq_c_w +
         (1 + visitN | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
# title: "dACC Activation Post-Stress Moderates \n Longitudinal Association between Life Stress and Anxiety"
sjPlot::plot_model(mlm_dacc_loss_prepost_mfq, type = "int", show.data = TRUE, colors = "Dark2", dot.size = 1, robust = TRUE, mdrt.values = "meansd", title = "", legend.title ="Residualized dACC \n Post-Pre Acute Stress")
save_plot("dACCxStress-Anxiety.png", fig = ggplot2::last_plot(), width = 12, height = 8, dpi = 300, theme = ggplot2::theme_get(), label.color = "black", label.size = 2.4, axis.textsize = 0.8, axis.titlesize = 0.75, legend.textsize = 0.6, legend.titlesize = 0.65, legend.itemsize = 0.5)

sim_slopes(mlm_dacc_loss_prepost_z_mfq, pred=aleq_c.L1_z, modx=dACC_prepost_resLOSS, jnplot=TRUE)

mlm_dacc_loss_prepost_z_shaps <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleq_c.L2_z +  aleq_c.L1_z * dACC_prepost_resLOSS  + shaps_c_z +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
summary(mlm_dacc_loss_prepost_z_shaps)
check_convergence(mlm_dacc_loss_prepost_z_shaps)
sjPlot::tab_model(
  mlm_dacc_loss_prepost_z_shaps, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234) # interaction holds 
sim_slopes(mlm_dacc_loss_prepost_z_shaps, pred=aleq_c.L1_z, modx=dACC_prepost_resLOSS, jnplot=TRUE)
```

# Table dACC
```{r}
sjPlot::tab_model(
  mlm_dacc_loss_prepost_z, 
  mlm_dacc_loss_prepost_z_mfq,
  mlm_dacc_loss_prepost_z_shaps,
  bootstrap = TRUE,
  iterations = 1000, 
  dv.labels = c("Anxiety Severity", "Anxiety Severity", "Anxiety Severity"), 
  seed = 1234,
  pred.labels = 
    c("Intercept",
      "Sex [F]",
      "Between-person Stress", 
      "Within-person Stress",
      "dACC",
      "Within-Person Stress* dACC",
      "Depression",
      "Anhedonia"
      ),
  file = "dACC_Main_Anxiety.doc")
```

#### Based on Reviewer Recommendation, testing interpersonal vs noninterersonal stress

```{r}
# dACC x within-person interpersonal stress
INTmlm_dacc_loss_prepost_z <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleqInt_c.L2_z +  aleqInt_c.L1_z * dACC_prepost_resLOSS  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
summary(INTmlm_dacc_loss_prepost_z)

# dACC x within-person non-interpersonal stress
NONINTmlm_dacc_loss_prepost_z <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleqNonInt_c.L2_z +  aleqNonInt_c.L1_z * dACC_prepost_resLOSS  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
summary(NONINTmlm_dacc_loss_prepost_z)

AIC(INTmlm_dacc_loss_prepost_z, NONINTmlm_dacc_loss_prepost_z)
BIC(INTmlm_dacc_loss_prepost_z, NONINTmlm_dacc_loss_prepost_z)


```

> Stronger for interpersonal compared to non-interpersonal stress

#### Based on Reviewer Recommendation, testing model with fixed effect of visit 

```{r}
visit_mlm_dacc_loss_prepost_z <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleq_c.L2_z +  aleq_c.L1_z * dACC_prepost_resLOSS  +
         visit_num_z + (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
summary(visit_mlm_dacc_loss_prepost_z)
```
> Adding fixed effect of visit number does not change the interaction between stress and ROI


### AntIns pre-post-stress loss
```{r}
mlm_antins_l_prepost_z <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleq_c.L2_z +  aleq_c.L1_z*AntIns_l_prepost_resLOSS  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
summary(mlm_antins_l_prepost_z) 
sjPlot::tab_model(
  mlm_antins_l_prepost_z, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234) #  No interaction  

mlm_antins_r_prepost_z <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleq_c.L2_z +  aleq_c.L1_z * AntIns_r_prepost_resLOSS  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
summary(mlm_antins_r_prepost_z) # aleq_c.L1_z:AntIns_r_prepost_resLOSS 
sjPlot::tab_model(
  mlm_antins_r_prepost_z, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234) # interaction

sim_slopes(mlm_antins_r_prepost_z, pred=aleq_c.L1_z, modx=AntIns_r_prepost_resLOSS, jnplot=TRUE, confint = TRUE)
probe_interaction(mlm_antins_r_prepost_z, aleq_c.L1_z, AntIns_r_prepost_resLOSS)

check_convergence(mlm_antins_r_prepost_z)
plot_model(mlm_antins_r_prepost_z, type = "diag") 
check_model(mlm_antins_r_prepost_z) 


####### Exploratory adjusting for longitudinal MFQ and SHAPS ####### 
mlm_antins_r_prepost_z_mfq <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleq_c.L2_z +  aleq_c.L1_z*AntIns_r_prepost_resLOSS  + mafq_c_w_z +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
summary(mlm_antins_r_prepost_z_mfq) 
check_convergence(mlm_antins_r_prepost_z_mfq)
sjPlot::tab_model(
  mlm_antins_r_prepost_z_mfq, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234) # interaction holds 

mlm_antins_r_prepost_mfq <- 
  lmer(
    masc_c_total ~ 
      Sex + aleq_c.L2 +  aleq_c.L1*AntIns_r_prepost_resLOSS  + mafq_c_w +
         (1 + visitN | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
sjPlot::plot_model(mlm_antins_r_prepost_mfq, type = "int", show.data = TRUE, colors = "Dark2", dot.size = 1, robust = TRUE, mdrt.values = "meansd", title = "R Anterior Insula Activation Post-Stress Moderates \n Longitudinal Association between Life Stress and Anxiety", legend.title ="Residualized R Ant Ins \n Post-Pre Acute Stress")
sim_slopes(mlm_antins_r_prepost_z_mfq, pred=aleq_c.L1_z, modx=AntIns_r_prepost_resLOSS, jnplot=TRUE)

mlm_antins_r_prepost_z_shaps <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleq_c.L2_z +  aleq_c.L1_z*AntIns_r_prepost_resLOSS + shaps_c_z +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
summary(mlm_antins_r_prepost_z_shaps)  
check_convergence(mlm_antins_r_prepost_z_shaps)
sjPlot::tab_model(
  mlm_antins_r_prepost_z_shaps, 
  bootstrap = TRUE,
  iterations = 1000, 
  seed = 1234) # interaction holds 

sim_slopes(mlm_antins_r_prepost_z_shaps, pred=aleq_c.L1_z, modx=AntIns_r_prepost_resLOSS, jnplot=TRUE)

```

#### Based on Reviewer Recommendation, testing interpersonal vs noninterersonal stress

```{r}
INTmlm_antins_r_prepost_z <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleqInt_c.L2_z +  aleqInt_c.L1_z * AntIns_r_prepost_resLOSS  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
summary(INTmlm_antins_r_prepost_z)

NONINTmlm_antins_r_prepost_z <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleqNonInt_c.L2_z +  aleqNonInt_c.L1_z * AntIns_r_prepost_resLOSS  +
         (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
summary(NONINTmlm_antins_r_prepost_z)
```

> Both interpersonal and non-interpersonal stress interact with the R ant insula to predict anxiety severity

#### Based on Reviewer Recommendation, testing model with fixed effect of visit 

```{r}
visit_mlm_antins_r_prepost_z <- 
  lmer(
    masc_c_total_z ~ 
      Sex + aleqInt_c.L2_z +  aleqInt_c.L1_z * AntIns_r_prepost_resLOSS  + visit_num_z + (1 + visit_num_z | subID), 
       data = pad_and_rois_merge_LOSS_z
  )
summary(visit_mlm_antins_r_prepost_z)
```
> Adding fixed effect of visit number does not change the interaction between stress and ROI

# Table R Ant Ins
```{r}
sjPlot::tab_model(
  mlm_antins_r_prepost_z, 
  mlm_antins_r_prepost_z_mfq,
  mlm_antins_r_prepost_z_shaps,
  bootstrap = TRUE,
  iterations = 1000, 
  dv.labels = c("Anxiety Severity", "Anxiety Severity", "Anxiety Severity"), 
  seed = 1234,
  pred.labels = 
    c("Intercept",
      "Sex [F]",
      "Between-person Stress", 
      "Within-person Stress",
      "R Anterior Insula",
      "Within-Person Stress* R Anterior Insula",
      "Depression",
      "Anhedonia"
      ),
  file = "RAntIns_Main_Anxiety.doc")
```


#### MC CORRECTION FOR LOSS

```{r}
# dACC, l Ant Ins, r Ant Ins
ploss <- c(0.008, 0.310, 0.002)
p.adjust(ploss, method = "fdr", n = length(ploss))
# 0.012 0.310 0.006
```

# Null Win Model Tables
```{r}
sjPlot::tab_model(
  mlm_NAcc_l_win_prepost_z, 
  mlm_NAcc_r_win_prepost_z,
  mlm_caudate_l_win_prepost_z,
  mlm_caudate_r_win_prepost_z,
  mlm_putamen_l_win_prepost_z,
  mlm_putamen_r_win_prepost_z,
  bootstrap = TRUE,
  iterations = 1000, 
  dv.labels = c("Anxiety Severity", "Anxiety Severity", "Anxiety Severity", "Anxiety Severity", "Anxiety Severity", "Anxiety Severity"), 
  seed = 1234,
  pred.labels = 
    c("Intercept",
      "Sex [F]",
      "Between-person Stress", 
      "Within-person Stress",
      "L NAcc",
      "Within-Person Stress* L NAcc",
      "R NAcc",
      "Within-Person Stress* R NAcc",
      "L Caudate",
      "Within-Person Stress* L Caudate",     
      "R Caudate",
      "Within-Person Stress* R Caudate",
      "L Putamen",
      "Within-Person Stress* L Putamen",  
      "R Putamen",
      "Within-Person Stress* R Putamen"          
      ),
  file = "Anxiety_NullResults_Win.doc")
```

# Null Loss Model Tables
```{r}
sjPlot::tab_model(
  mlm_antins_l_prepost_z, 
  bootstrap = TRUE,
  iterations = 1000, 
  dv.labels = c("Anxiety Severity"), 
  seed = 1234,
  pred.labels = 
    c("Intercept",
      "Sex [F]",
      "Between-person Stress", 
      "Within-person Stress",
      "L Anterior Insula",
      "Within-Person Stress* L Anterior Insula"
      ),
  file = "Anxiety_NullResults_Loss.doc")
```
